+++ daeng/src/components/review/TotalReviewForm.jsx
@@ -190,8 +190,8 @@ const TotalReviewForm = () => {
           return (
             <div
               key={review.reviewId}
-              ref={index === reviews.length - 1 ? observeLastItem : null}
-            >
+              ref={index === reviews.length - 1 ? observeLastItem : null}//observeLastItem을 통한 마지막 요소 감지
+            > 
               <ReviewUserContainer>
                 <UserPhoto
                   src={review.petImg || reviewDefaultImg}
